requirements.txt
aiogram==3.7.0
aiosqlite==0.20.0
reportlab==4.1.0
Pillow==10.3.0
python-dotenv==1.0.1
aiofiles==23.2.1
